+++ apps/owner/src/components/BottomTabBar/OwnerBottomTabBar.tsx
@@ -22,6 +22,7 @@ const ownerTabs = [
 const ownerHideTabbarRoutes = [
   '/example',
   '/review',
+  '/signup',
   // 숨기고 싶은 tabbar route 추가
 ];
