+++ packages/design-system/components/TypeOneButton/TypeOneButton.tsx
@@ -20,7 +20,7 @@ const TypeOneButton = ({ text, onClick, color, className = '', type, disabled }:
     <div className='absolute bottom-0 w-full max-w-[480px]'>
       <button
         type={type ? type : 'button'}
-        disabled={disabled ? disabled : true}
+        disabled={disabled ? disabled : false}
         aria-label={text}
         className={`min-h-[60px] w-full cursor-pointer px-4 py-2 text-sub_h1 ${color} ${textColor} ${className}`}
         onClick={onClick}
